requirements.txt
requests-2.31.0
environs-10.3.0
terminaltables-3.1.10
